Move the Shell struct and impl into their own file
